+++ src/argsParser.ts
@@ -40,8 +40,8 @@ function processOptions(
       case '--excludeDeclarationFiles':
         newOptions.excludeDeclarationFiles = true;
         break;
-      case '--enableSearchNamespaces':
-        newOptions.enableSearchNamespaces = true;
+      case '--searchNamespaces':
+        newOptions.searchNamespaces = true;
         break;
       case '--showLineNumber':
         newOptions.showLineNumber = true;