+++ src/types.ts
@@ -41,6 +41,6 @@ export interface ExtraCommandLineOptions {
   exitWithCount?: boolean;
   excludeDeclarationFiles?: boolean;
   pathsToIgnore?: string[];
-  enableSearchNamespaces?: boolean;
+  searchNamespaces?: boolean;
   showLineNumber?: boolean;
 }