+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased] - ReleaseDate
 
+### Added
+- Handle exports from within a namespace
+
 ### Changed
 - (Internal) Update dependency TypeScript to 3.7.3
 - (Internal) Simplify some logic, using the new optional chaining operator (?.)