+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [Unreleased] - ReleaseDate
 
 ### Added
-- Handle exports from within a namespace
+
+- Handle exports from within a namespace. Disabled when option --disableSearchNamespaces is given. Note: there is a performance cost when this is turned on (about 5x slower on large codebases).
 
 ### Changed
 - (Internal) Update dependency TypeScript to 3.7.3